project name modification
