+++ src/main/java/com/flaviolcord/user/registry/UserRegistryApplication.java
@@ -1,0 +1,15 @@
+package com.flaviolcord.user.registry;
+
+import com.flaviolcord.user.registry.infrastructure.config.UserProperties;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+
+@SpringBootApplication
+@EnableConfigurationProperties(UserProperties.class)
+public class UserRegistryApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UserRegistryApplication.class, args);
+	}
+}
