--- src/main/java/com/flaviolcord/user/registry/api/UserRegistryApiApplication.java
@@ -1,13 +1,0 @@
-package com.flaviolcord.user.registry.api;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UserRegistryApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UserRegistryApiApplication.class, args);
-	}
-
-}
