+++ src/main/java/com/flaviolcord/user/registry/domain/model/User.java
@@ -1,29 +1,41 @@
-package com.flaviolcord.user.registry.api.dto;
+package com.flaviolcord.user.registry.domain.model;
 
-import jakarta.validation.constraints.*;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Past;
+import jakarta.validation.constraints.Pattern;
+import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.Setter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 import java.time.LocalDate;
 
-@Getter @Setter @NoArgsConstructor
-public class UserDTO {
+@Entity
+@Table(name = "users")
+@Getter @Setter @NoArgsConstructor @AllArgsConstructor
+public class User {
 
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @NotNull(message = "Username is required")
+    @NotNull
     @Size(min = 3, max = 50, message = "Username must be between 3 and 50 characters")
     private String username;
 
-    @NotNull(message = "Birthdate is required")
+    @NotNull
     @Past(message = "Birthdate must be in the past")
     private LocalDate birthdate;
 
-    @NotNull(message = "Country of residence is required")
+    @NotNull
+    @Column(name = "country_of_residence")
     @Pattern(regexp = "France", message = "Only French residents are allowed")
     private String countryOfResidence;
 
+    @Column(name = "phone_number")
     @Size(min = 10, max = 15, message = "Phone number must be between 10 and 15 characters")
     private String phoneNumber; // Optional
