+++ src/main/java/com/flaviolcord/user/registry/infrastructure/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.flaviolcord.user.registry.api.exception;
+package com.flaviolcord.user.registry.infrastructure.exception;
 
 public class UserNotFoundException extends RuntimeException {
