+++ src/main/java/com/flaviolcord/user/registry/infrastructure/exception/UserRegistrationException.java
@@ -1,10 +1,6 @@
-package com.flaviolcord.user.registry.api.exception;
+package com.flaviolcord.user.registry.infrastructure.exception;
 
 public class UserRegistrationException extends RuntimeException {
-
-    public UserRegistrationException(String message) {
-        super(message);
-    }
 
     public UserRegistrationException(String message, Throwable cause) {
         super(message, cause);
