+++ src/main/java/com/flaviolcord/user/registry/infrastructure/exception/ValidationException.java
@@ -1,4 +1,4 @@
-package com.flaviolcord.user.registry.api.exception;
+package com.flaviolcord.user.registry.infrastructure.exception;
 
 public class ValidationException extends RuntimeException {
