+++ src/main/resources/application.properties
@@ -1,3 +1,14 @@
-spring.application.name=User Registry API
-user.allowed.country=France
-user.min.age=18
+spring.application.name=User Registry
+user.allowedCountry=France
+user.minAge=18
+
+# In-memory database
+spring.datasource.url=jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=password
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.hibernate.ddl-auto=update
+spring.h2.console.enabled=true
+
+spring.jpa.properties.hibernate.transaction.jta.platform=org.hibernate.engine.transaction.jta.platform.internal.NoJtaPlatform
