+++ src/test/java/com/flaviolcord/user/registry/UserRegistryApiApplicationTests.java
@@ -1,4 +1,4 @@
-package com.flaviolcord.user.registry.api;
+package com.flaviolcord.user.registry;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -9,5 +9,4 @@ class UserRegistryApiApplicationTests {
 	@Test
 	void contextLoads() {
 	}
-
 }
